Move review submission components to their own folder
