+++ site/gatsby-site/src/components/submissions/SubmissionList.js
@@ -1,11 +1,9 @@
 import React from 'react';
 import ListGroup from 'react-bootstrap/ListGroup';
 import { Spinner } from 'react-bootstrap';
-
 import Link from 'components/ui/Link';
-import ReportedIncident from 'components/ReportedIncident';
-
-import { FIND_SUBMISSIONS } from '../graphql/submissions';
+import SubmissionReview from 'components/submissions/SubmissionReview';
+import { FIND_SUBMISSIONS } from '../../graphql/submissions';
 import { useQuery } from '@apollo/client';
 
 const SubmissionList = () => {
@@ -31,9 +29,9 @@ const SubmissionList = () => {
             .sort(
               (a, b) => new Date(a.date_submitted).getTime() - new Date(b.date_submitted).getTime()
             )
-            .map((node) => (
-              <ListGroup.Item key={node._id} className="m-0 p-0">
-                <ReportedIncident incident={node} />
+            .map((submission) => (
+              <ListGroup.Item key={submission._id} className="m-0 p-0">
+                <SubmissionReview incident={submission} />
               </ListGroup.Item>
             ))}
       </ListGroup>
